Separate config and command into separate modules
